refactor: Name of hooks
